Rename to follow Nette naming conventions
- Rename interface from LinkGeneratorInterface to LinkGenerator (Nette uses plain nouns for interfaces)
- Rename class from LinkGenerator to DefaultLinkGenerator
- Update all references in Presenter, ApplicationExtension, LinkBaseNode
- Update all existing tests to use DefaultLinkGenerator for instantiation
- Update static method calls: DefaultLinkGenerator::parseDestination(), DefaultLinkGenerator::applyBase()
- Update Latte bridge compiled template output assertions
