+++ src/Bridges/ApplicationDI/ApplicationExtension.php
@@ -12,7 +12,7 @@ namespace Nette\Bridges\ApplicationDI;
 use Composer\Autoload\ClassLoader;
 use Nette;
 use Nette\Application\Attributes;
-use Nette\Application\LinkGeneratorInterface;
+use Nette\Application\LinkGenerator;
 use Nette\Application\UI;
 use Nette\DI\Definitions;
 use Nette\Schema\Expect;
@@ -115,8 +115,8 @@ final class ApplicationExtension extends Nette\DI\CompilerExtension
 		}
 
 		$builder->addDefinition($this->prefix('linkGenerator'))
-			->setType(LinkGeneratorInterface::class)
-			->setFactory(Nette\Application\LinkGenerator::class, [
+			->setType(LinkGenerator::class)
+			->setFactory(Nette\Application\DefaultLinkGenerator::class, [
 				1 => new Definitions\Statement([new Definitions\Statement('@Nette\Http\IRequest::getUrl'), 'withoutUserInfo']),
 			]);
