+++ tests/Bridges.DI/ApplicationExtension.basic.phpt
@@ -25,6 +25,6 @@ test('', function () {
 	$container = new Container1;
 	Assert::type(Nette\Application\Application::class, $container->getService('application'));
 	Assert::type(Nette\Application\PresenterFactory::class, $container->getService('nette.presenterFactory'));
-	Assert::type(Nette\Application\LinkGenerator::class, $container->getService('application.linkGenerator'));
-	Assert::type(Nette\Application\LinkGeneratorInterface::class, $container->getByType(Nette\Application\LinkGeneratorInterface::class));
+	Assert::type(Nette\Application\DefaultLinkGenerator::class, $container->getService('application.linkGenerator'));
+	Assert::type(Nette\Application\LinkGenerator::class, $container->getByType(Nette\Application\LinkGenerator::class));
 });
